+++ src/platform/SIMULATOR/sitl.c
@@ -186,15 +186,15 @@ void updateState(const fdm_packet* pkt)
 #endif
 
 #if defined(USE_VIRTUAL_GPS)
-    float longitude, latitude, altitude, speed, speed3D, course;
-    longitude = (float)pkt->position_xyz[0];
-    latitude = (float)pkt->position_xyz[1];
-    altitude = (float)pkt->position_xyz[2];
-    speed = (float)sqrt(pkt->velocity_xyz[0] * pkt->velocity_xyz[0] + pkt->velocity_xyz[1] * pkt->velocity_xyz[1]);
-    speed3D = (float)sqrt(pkt->velocity_xyz[0] * pkt->velocity_xyz[0] + pkt->velocity_xyz[1] * pkt->velocity_xyz[1] + pkt->velocity_xyz[2] * pkt->velocity_xyz[2]);
-    course = (float)(atan2(pkt->velocity_xyz[0], pkt->velocity_xyz[1]) * RAD2DEG);
-    if (course < 0.0f)
-        course += 360.0f;
+    double longitude, latitude, altitude, speed, speed3D, course;
+    longitude = pkt->position_xyz[0];
+    latitude = pkt->position_xyz[1];
+    altitude = pkt->position_xyz[2];
+    speed = sqrt(pkt->velocity_xyz[0] * pkt->velocity_xyz[0] + pkt->velocity_xyz[1] * pkt->velocity_xyz[1]);
+    speed3D = sqrt(pkt->velocity_xyz[0] * pkt->velocity_xyz[0] + pkt->velocity_xyz[1] * pkt->velocity_xyz[1] + pkt->velocity_xyz[2] * pkt->velocity_xyz[2]);
+    course = atan2(pkt->velocity_xyz[0], pkt->velocity_xyz[1]) * RAD2DEG;
+    if (course < 0.0)
+        course += 360.0;
     setVirtualGPS(latitude, longitude, altitude, speed, speed3D, course);
 #endif
